requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-pinecone
langchain-huggingface
pypdf2
sentence-transformers
pinecone
pyreadline3       